Move the fingerprint model fixture in with its siblings
test_trained_model.py hand-wrote the three recipe YAMLs that conftest already
writes for the other model directories, so the test module carried 55 lines of
literals and its own file handles.

Generalize _write_recipe_components with name and feat, defaulting feat to the
NullFeaturizer spec so the existing fixtures are untouched, and move
fingerprint_model_dir next to the two it belongs with. All four model-directory
fixtures now live in one place and the test module opens no files.
